REFACTOR: Begin new submodule for (co)limits in general categories.
