+++ src/categorical_algebra/ShapeDiagrams.jl
@@ -1,23 +1,17 @@
 """ Diagrams of a given shape.
 """
 module ShapeDiagrams
-export AbstractFunctor, AbstractLaxator, LaxMonoidalFunctor,
-  Span, Cone, Cospan, DecoratedCospan, Cocone, Diagram, left, right, leg, nlegs, apex, base,
+export Span, Cospan, Diagram, apex, base, left, right,
+  DecoratedCospan, AbstractFunctor, AbstractLaxator, LaxMonoidalFunctor,
   decorator, decoration, undecorate
+
+using AutoHashEquals
 
 using ...Theories: dom, codom
 
-abstract type AbstractFunctor end
-abstract type AbstractLaxator end
-
-struct LaxMonoidalFunctor{Ftr <: AbstractFunctor, Lxr <: AbstractLaxator} <: AbstractFunctor
-  F::Ftr
-  L::Lxr
-end
-
 """ Span of morphisms in a category.
 """
-struct Span{Left,Right}
+@auto_hash_equals struct Span{Left,Right}
   left::Left
   right::Right
 
@@ -33,28 +27,9 @@ apex(span::Span) = dom(span.left) # == dom(span.right)
 left(span::Span) = span.left
 right(span::Span) = span.right
 
-""" Cone of morphisms in a category.
-"""
-struct Cone{S,T}
-  apex::S # We store this separately because legs might be empty
-  legs::Vector{T}
-
-  function Cone(apex::S,legs::Vector{T}, strict::Bool=true) where {S,T}
-    if strict && !all(dom(leg) == apex for leg in legs)
-      error("Domain of legs in cone do not match")
-    end
-    new{S,T}(apex,legs)
-  end
-end
-
-apex(cone::Cone) = cone.apex
-leg(cone::Cone,n) = cone.legs[n]
-legs(cone::Cone) = cone.legs
-nlegs(cone::Cone) = length(cone.legs)
-
 """ Cospan of morphisms in a category.
 """
-struct Cospan{Left,Right}
+@auto_hash_equals struct Cospan{Left,Right}
   left::Left
   right::Right
 
@@ -70,6 +45,14 @@ base(cospan::Cospan) = codom(cospan.left) # == codom(cospan.right)
 left(cospan::Cospan) = cospan.left
 right(cospan::Cospan) = cospan.right
 
+# FIXME: 
+abstract type AbstractFunctor end
+abstract type AbstractLaxator end
+
+struct LaxMonoidalFunctor{Ftr <: AbstractFunctor, Lxr <: AbstractLaxator} <: AbstractFunctor
+  F::Ftr
+  L::Lxr
+end
 
 """ Decorate Cospan of morphisms for representing open networks.
 """
@@ -86,24 +69,6 @@ base(m::DecoratedCospan) = base(m.cospan)
 left(m::DecoratedCospan) = left(m.cospan)
 right(m::DecoratedCospan) = right(m.cospan)
 
-""" Cocone of morphisms in a category.
-"""
-struct Cocone{S,T}
-  base::S # We store this separately because legs might be empty
-  legs::Vector{T}
-
-  function Cocone(base::S,legs::Vector{T}, strict::Bool=true) where {S,T}
-    if strict && !all(codom(leg) == base for leg in legs)
-      error("Codomain of legs in cocone do not match")
-    end
-    new{S,T}(base,legs)
-  end
-end
-
-base(cocone::Cocone) = cocone.base
-leg(cocone::Cocone,n) = cocone.legs[n]
-nlegs(cocone::Cocone) = length(cocone.legs)
-
 struct Diagram{Ob,Hom}
   obs::Vector{Ob}
   homs::Vector{Tuple{Int64,Int64,Hom}}
@@ -116,6 +81,5 @@ struct Diagram{Ob,Hom}
     new{Ob,Hom}(obs,homs)
   end
 end
-
 
 end
